Added more meta data to the K2 Classic style.
Added preliminary support for 'Content Width:' in styles, which will be used for...

Added preliminary support for resizing ('480px × 295px'-only) YouTube embeds to content width. Would like to expand this to support flickr and other video services as well at some point. For now it needs to become aware of the dominating style's content width and refactor the height based on that.

git-svn-id: http://kaytwo.googlecode.com/svn/trunk@1014 0da3f984-a319-0410-8551-939afe8cdb0b